+++ src/test/java/org/javierperis/pricing/PriceCalculationInteractorTest.java
@@ -19,19 +19,19 @@ import static org.mockito.Mockito.*;
 @ExtendWith(MockitoExtension.class)
 public class PriceCalculationInteractorTest {
 
-    private static final Long ZARA_BRAND_ID = 1L;
+    private static final Integer ZARA_BRAND_ID = 1;
     private static final Long PRODUCT_ID = 35455L;
     private static final LocalDateTime firstStartDate = LocalDate.of(2020, 6, 14)
             .atTime(0, 0);
     private static final LocalDateTime firstEndDate = LocalDate.of(2020, 12, 31)
             .atTime(23, 59);
-    private static final Long FIRST_PRICE_LIST = 1L;
+    private static final Integer FIRST_PRICE_LIST = 1;
     private static final Double FIRST_PRICE = 35.50D;
     private static final LocalDateTime secondStartDate = LocalDate.of(2020, 6, 14)
             .atTime(15, 0);
     private static final LocalDateTime secondEndDate = LocalDate.of(2020, 6, 14)
             .atTime(18, 30);
-    private static final Long SECOND_PRICE_LIST = 2L;
+    private static final Integer SECOND_PRICE_LIST = 2;
     private static final Double SECOND_PRICE = 25.45D;
 
     @Mock
@@ -50,9 +50,9 @@ public class PriceCalculationInteractorTest {
     @Test
     public void givenJustOnePriceInRange_whenRequest_ThenReturnThatPrice() {
         final LocalDateTime localDateTime = LocalDate.of(2020, 6, 16).atTime(21, 0);
-        final PriceRequestModel priceRequestModel = new PriceRequestModel(1L, 35455L, localDateTime);
+        final PriceRequestModel priceRequestModel = new PriceRequestModel(1, 35455L, localDateTime);
         final List<PriceDsResponseModel> prices = List.of(createPriceDsResponse(FIRST_PRICE_LIST,
-                firstStartDate, firstEndDate, 0L, FIRST_PRICE));
+                firstStartDate, firstEndDate, 0, FIRST_PRICE));
         final PriceResponseModel expectedPriceResponseModel = createPriceResponseModel(FIRST_PRICE_LIST,
                 firstStartDate, firstEndDate, FIRST_PRICE);
         when(priceCalculationDsGateway.getPrices(any())).thenReturn(prices);
@@ -65,10 +65,10 @@ public class PriceCalculationInteractorTest {
     @Test
     public void givenSomePricesInRange_whenRequest_ThenReturnPriceWithHigherPriority() {
         final LocalDateTime localDateTime = LocalDate.of(2020, 6, 14).atTime(16, 0);
-        final PriceRequestModel priceRequestModel = new PriceRequestModel(1L, 35455L, localDateTime);
+        final PriceRequestModel priceRequestModel = new PriceRequestModel(1, 35455L, localDateTime);
         final List<PriceDsResponseModel> prices = List.of(
-                createPriceDsResponse(FIRST_PRICE_LIST, firstStartDate, firstEndDate, 0L, FIRST_PRICE),
-                createPriceDsResponse(SECOND_PRICE_LIST, secondStartDate, secondEndDate, 1L, SECOND_PRICE));
+                createPriceDsResponse(FIRST_PRICE_LIST, firstStartDate, firstEndDate, 0, FIRST_PRICE),
+                createPriceDsResponse(SECOND_PRICE_LIST, secondStartDate, secondEndDate, 1, SECOND_PRICE));
         final PriceResponseModel expectedPriceResponseModel = createPriceResponseModel(SECOND_PRICE_LIST,
                 secondStartDate, secondEndDate, SECOND_PRICE);
         when(priceCalculationDsGateway.getPrices(any())).thenReturn(prices);
@@ -80,7 +80,7 @@ public class PriceCalculationInteractorTest {
     @Test
     public void givenNoPriceInRage_whenRequest_ThenReturnError() {
         final LocalDateTime localDateTime = LocalDate.of(2020, 6, 14).atTime(16, 0);
-        final PriceRequestModel priceRequestModel = new PriceRequestModel(1L, 35455L, localDateTime);
+        final PriceRequestModel priceRequestModel = new PriceRequestModel(1, 35455L, localDateTime);
 
         when(priceCalculationDsGateway.getPrices(any())).thenReturn(new ArrayList<>());
 
@@ -89,14 +89,14 @@ public class PriceCalculationInteractorTest {
                 .prepareFailView("There's no price for the product and date specified");
     }
 
-    private PriceDsResponseModel createPriceDsResponse(Long priceList, LocalDateTime startDateForApplicablePrice,
-                                                       LocalDateTime endDateForApplicablePrice, Long priority,
+    private PriceDsResponseModel createPriceDsResponse(Integer priceList, LocalDateTime startDateForApplicablePrice,
+                                                       LocalDateTime endDateForApplicablePrice, Integer priority,
                                                        double expectedPrice) {
         return new PriceDsResponseModel(PRODUCT_ID, ZARA_BRAND_ID, priceList,
                 startDateForApplicablePrice, endDateForApplicablePrice, priority, expectedPrice, "EUR");
     }
 
-    private PriceResponseModel createPriceResponseModel(Long priceList, LocalDateTime startDateForPrice,
+    private PriceResponseModel createPriceResponseModel(Integer priceList, LocalDateTime startDateForPrice,
                                                         LocalDateTime endDateForPrice, double expectedPrice) {
         return new PriceResponseModel(PRODUCT_ID, ZARA_BRAND_ID, priceList, startDateForPrice, endDateForPrice,
                 expectedPrice, Currency.getInstance("EUR"));
